Deal with inconsistently firing onpopstate
Some browsers don't support pushState yet and so have no state to pop and fire,
while others happily fire onpopstate as soon as the page loads. When the
pageload fires, the state object is null, where if the user hit back from a
pushState, the state object is present. By detecting that, the pageload event
can be ignored and the first load can be manual to accomodate any browser.

The next tricky bit is that if we roll all the way back to the initial pageload
state, it comes in as null again, but this time the pop should happen. So what's
another one-way state boolean among friends?