fix urls for 'clever' clients that load images
Singleton tweet URLs end with an image, causing some clients to try to load them
inline. URI escaping functions happily leave behind periods and don't solve the
problem, so kludge it by replacing a period near the end with an underscore.
Only do this to periods within 5 chars of the end to avoid messing with some
other part of the string, and do the same length check to accept a URL with a
period still there if it comes in.